# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Live Restaurant Queue & Pre-Order Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /restaurants/{restaurantId}/queue:
    get:
      operationId: getRestaurantQueue
      tags: [queue]
      summary: Get all queue entries for a restaurant (restaurant dashboard)
      parameters:
        - name: restaurantId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/QueueEntry"
    post:
      operationId: joinQueue
      tags: [queue]
      summary: Customer joins the queue (no login required)
      parameters:
        - name: restaurantId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JoinQueueInput"
      responses:
        "201":
          description: Joined queue
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueueEntry"
        "400":
          description: Queue is closed or invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /queue/{sessionToken}:
    get:
      operationId: getQueueEntry
      tags: [queue]
      summary: Get a customer's queue entry by session token (polling endpoint)
      parameters:
        - name: sessionToken
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueueEntryWithContext"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /queue/{sessionToken}/leave:
    post:
      operationId: leaveQueue
      tags: [queue]
      summary: Customer leaves the queue
      parameters:
        - name: sessionToken
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /queue-entries/{entryId}/call:
    post:
      operationId: callQueueEntry
      tags: [queue]
      summary: Restaurant calls a customer (marks as called)
      parameters:
        - name: entryId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueueEntry"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /queue-entries/{entryId}/seat:
    post:
      operationId: seatQueueEntry
      tags: [queue]
      summary: Restaurant seats a customer
      parameters:
        - name: entryId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueueEntry"
  /queue-entries/{entryId}/remove:
    post:
      operationId: removeQueueEntry
      tags: [queue]
      summary: Restaurant removes a no-show customer
      parameters:
        - name: entryId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueueEntry"

  # ─── Menu ─────────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean

    # ─── Auth ──────────────────────────────────────────────────────────────────
...
    RestaurantPublic:
      type: object
      required: [id, name, address, cuisineType, avgWaitPerTableMins, queueOpen, queueLength, estimatedWaitMins]
      properties:
        id:
          type: integer
        name:
          type: string
        address:
          type: string
        cuisineType:
          type: string
        photo:
          type: ["string", "null"]
        avgWaitPerTableMins:
          type: integer
        queueOpen:
          type: boolean
        preOrderThreshold:
          type: integer
        queueLength:
          type: integer
        estimatedWaitMins:
          type: integer
...
    JoinQueueInput:
      type: object
      required: [customerName, customerPhone, partySize]
      properties:
        customerName:
          type: string
        customerPhone:
          type: string
        partySize:
          type: integer
          minimum: 1
    QueueEntry:
      type: object
      required: [id, restaurantId, queueNumber, customerName, customerPhone, partySize, status, joinedAt, sessionToken]
      properties:
        id:
          type: integer
        restaurantId:
          type: integer
        queueNumber:
          type: integer
        customerName:
          type: string
        customerPhone:
          type: string
        partySize:
          type: integer
        status:
          type: string
          enum: [waiting, called, seated, removed]
        joinedAt:
          type: string
        sessionToken:
          type: string
        orderId:
          type: ["integer", "null"]
    QueueEntryWithContext:
      type: object
      required: [entry, position, estimatedWaitMins, restaurant, canPreOrder]
      properties:
        entry:
          $ref: "#/components/schemas/QueueEntry"
        position:
          type: integer
        estimatedWaitMins:
          type: integer
        restaurant:
          $ref: "#/components/schemas/RestaurantPublic"
        canPreOrder:
          type: boolean
        order:
          $ref: "#/components/schemas/Order"

    # ─── Menu ──────────────────────────────────────────────────────────────────
...
    OrderItem:
      type: object
      required: [menuItemId, name, qty, price]
      properties:
        menuItemId:
          type: integer
        name:
          type: string
        qty:
          type: integer
        price:
          type: number
...
    Order:
      type: object
      required: [id, queueEntryId, restaurantId, items, status, totalAmount, paymentStatus, createdAt]
      properties:
        id:
          type: integer
        queueEntryId:
          type: integer
        restaurantId:
          type: integer
        customerName:
          type: string
        queueNumber:
          type: integer
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        status:
          type: string
          enum: [placed, preparing, ready, served]
        totalAmount:
          type: number
        paymentStatus:
          type: string
          enum: [pending, paid_cash, paid_online]
        createdAt:
          type: string
...