# lib/api-zod/src/generated/types/joinQueueInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Live Restaurant Queue & Pre-Order Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface JoinQueueInput {
  customerName: string;
  customerPhone: string;
  /** @minimum 1 */
  partySize: number;
}