# lib/api-zod/src/generated/types/queueEntry.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Live Restaurant Queue & Pre-Order Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { QueueEntryStatus } from './queueEntryStatus';

export interface QueueEntry {
  id: number;
  restaurantId: number;
  queueNumber: number;
  customerName: string;
  customerPhone: string;
  partySize: number;
  status: QueueEntryStatus;
  joinedAt: string;
  sessionToken: string;
  /** @nullable */
  orderId?: number | null;
}